Algorithm & Time complexity
𝑺𝒐𝒓𝒕 𝒄𝒉𝒂𝒓𝒂𝒄𝒉𝒕𝒆𝒓𝒔 𝒃𝒚 𝒇𝒓𝒆𝒒𝒖𝒆𝒏𝒄𝒚 🕐|| 𝑶(𝑵𝒍𝒐𝒈(𝑵)) 🧠 || 𝒒𝒖𝒊𝒄𝒌𝒔𝒐𝒓𝒕 🏃

Given a string s, sort it in decreasing order based on the frequency of the characters. The frequency of a character is the number of times it appears in the string.

Return the sorted string. If there are multiple answers, return any of them.

step 1:
create a hashmap and call it F when F[ c ] = the frequency of c in s

than well sort the array with first priority to F values than th esecond priority will be alphabetic so the carachters will not mix with each other for example without sorting using alphabetic order if F[ a ] = 2 and F[ b ] = 2 than we can get the result "abab" but by also sorting alphabeticly we get "aabb"

Time complexity : O(N log(N))

O(N) : building F
O(N * log(N)) : sorting 

Space complexity: O(N):
F hashmap : O(N)